LPD-34232 Rename fixture following pattern
